refactored code to extract column rendering methods
